refactor: change the type `Long` to `Int` for `duration` field
EPMDJ-10980

+++ autotest-agent/src/jvmMain/kotlin/com/epam/drill/agent/test/sending/TestLaunchPayload.kt
@@ -23,6 +23,6 @@ data class TestLaunchPayload(
     val testLaunchId: String,
     val testDefinitionId: String,
     val result: TestResult,
-    val duration: Long?,
+    val duration: Int?,
     val details: TestDefinitionPayload,
 )
